Add class `Room`, tests  and `Room#connect_to`
`Room#connect_to` is a method that connects one room to the other while
making sure that the connection is both ways and takes care of absolute
and relative orientation. I will write more about that in the docs so
read it there.

`Room#connect_to` method is created to be used exclusively by the
dungeon generator.

Since at this point there are no rooms implemented for the testing I
have put a bunch of them in the test/test_helper.rb along with the
helper methods defined locally in the `Room` class to provide a way to
test weather the rooms are properly connected between each other.