Add `Dungeon` class,tests and `Dungeon::generate`
`Dungeon::generate` is a public method that returns an array of rooms
that are all properly connected, it is the only one that should be
called with intention of generating the dungeon.

When the dungeon is generated the hero will be in the first room. Hero
movement from there onwards should be handled mostly trough
`Room#interact`.