Add `Room#hero_here?`, orb event and hero movement
This commit adds hero movement trough the rooms when the player exits
the room. It also adds a possibility of an healing orb event occurring
where the Hero (if damaged) has a chance to be healed for 1 hit point.

After the heal the orb breaks, because the orb is in the hallway between
two rooms it is important that both rooms are updated with this
information. Both this information and hero instance is transfered via
modifiers parameter to the next room, not the most elegant solution but
it will do for now.

Because the idea is never to break narrative immersion with numbers and
stats the heal reports as "two more hits before I perish" for example.
For now this is rather crudely done and I will refine it if it get
necessary to do so. So far the idea is for a player to keep aware of the
Hero's hit points since there will be no interface or helper methods for
this.