Add `SaveEvent` class and save event
This commit adds save prompt when the Hero exits the room. It currently
saves before the actual exiting of the room. I don't like the current
way it is done (stuffing description into modifiers, in same way as
stuffing hero reference in there as well).

For now I am just trying just to make it work, and if I am able and have
the time I will try to improve it.