Add `Book` class, tests and YML files
This commit adds a `Book` class which reads descriptions from the YML
files that are used to describe a room, or an item or an event etc. It
basically puts all of the text that will be outputted to the user in a
separate file, so that it doesn't mix up with the logic. More about in
in documentation.

`Book` class accepts any class method that it then runs trough
`Book::method_missing` which returns a class variable that has the hash
with the descriptions or if the variable does not exists, creates it and
assigns it the hash from the YML.

While it usefulness might be debated, at least I learned a little
metaprogramming.