Add `EntranceRoom` class, tests and text data
`EntranceRoom` is the where the Hero will appear at the start of the
game. I used this class to get the handle of inheritance in Ruby and
testing out the idea of having other room inherit `Room` and for the
most part I like it.

The parts I don't like is using the `modifiers` parameter of
`Room#interact` to store a bunch of different stuff used for flow
control but since I don't have a better idea at the time it will have to
do.