Add `Satchel, Item, AmuletItem, ShinyItem` classes
`Satchel` class is used for storage of obtained items and is build
in the similar manner as the `Room` class, meaning that it uses
`interact` to get the input from the command line.

It is different that it is activated by the command `satchel`
(or its synonyms). The hero picks up the satchel in the hallway
before the first room.

`Item` provides a set of methods that should be implemented by the
subclasses and it holds the typical attributes that every item should
have. In the same manner as the `Room` class it also holds a list
of classes that have inherited the class.

`AmuletItem` is the "main quest item". It is meant to store shiny
pieces that hero finds in the rooms and when combined with the
amulet is allows access to the final `EndRoom`

In this commit just the basic structure has been done which allows
interaction with the satchel via `Satchel#inspect` that is called
trough interact. The picking up and using the item from the
inventory should be done in the future commits.

This commit also features some grammar and if/else logic fixes.