Add `CLI, Thesaurus` classes, fix load bug
The bug happens when one attempts to load the game and gets stuck in the
save event, since the save occurs within the event class that has no
knowledge of the room. I fixed this by adding a check for a modifier
that will change the `current_even` flag. The solution is not the most
elegant one but if will do for now.

`Thesaurus` has one task, and that is to check the supplied (via
`Thesaurus.add_words`) hashes from the given text and return the
keywords of the words that match. I does not know about `Room` class
from which it gets the keywords and synonyms nor the `CLI` which uses
it. However, it requires a specific structure which is the same one
that `Room` uses it which means I will have to create a sort of data
transfer class to keep things as decoupled as I can.

`CLI` task is to provide proper formated output for the console and to
hold the main game loop, this should probably not be in the same class
so I might look into it in the future.

This is one big messy commit, but also a shift in perspective.

I decided to concentrate on the interface first and making things just
work, while leaving the specifics for later.

While developing for the aspect of modules that will later on come
together was very helpful for the dungeon generator the more I come
with the user interaction the more I notice that this approach is
better because it is more palpable for me.