Improve `Satchel#inspect`
Improved `Satchel#inspect` so that there are no more hard-coded values
like `:amulet` now it is dynamic, so that there is no need to type an
if-else clause for every item there.

The other idea I had is that inspect should display more descriptions if
more are provided. For example user types in `inspect amulet, shiny from
the satchel` and it would return both descriptions instead of just the
one like now. For now I will keep it simple.