Add `Satchel#put`
Added ability to put stuff in the satchel. Added more methods to
simple_room where I placed the first shiny item. I intentionally
kept the saved game pointing at the room to make it easier to test.

The more I work on this the more it gets bloated the bad design choices
start to show. The problem is that I simply don't know how to do it
better so it is mostly patching stuff to make it fit and work in this
way. I am glad I can make it work though.

I'll keep doing it like so until I get some better idea of how I should
handle it.