Add `Satchel#use`
Added ability to use items from the inventory. Aside from that
I also added an ability for a room to react when item is picked up
and used though that still needs work as well as the `Satchel#use`.

`Satchel#use` basically works in this pattern:
"Use(verb) amulet(noun) from the satchel(noun) on the wall(noun)."

So beside the verb "use" and noun "satchel" it requires both the
key item and the place where to use it. I might modify `Satchel#put`
to use similar pattern if the need arises.