Add `Satchel#combine`
Added the ability to combine items within the satchel. Pattern is:
Combine(verb) amulet(noun) and shiny(noun) piece(noun) in the
stash(noun).

Combining items raises the quality of the resulting item. If the
two items are the same it raises item's quantity, and if a new item
is created it has a default quality and quantity. Combinations are
based on the `@recipes` in the `Satchel` class.