Complete `SimpleRoom` story, add verbs and nouns.
For the most part the flags, nouns and verbs as well as the story for
the `SimpleRoom` is done.

Every room, or at least some rooms should have a different way of
obtaining the shiny, here a Hero has to try out the basic commands and
use the hints from the responses to find the right verbs and nouns to
use.