Add `Event` class, improve `SaveEvent` class
This commit adds `Event` class to serve as a parent class for all
events, and moves the `SaveEvent` class to the `events/` folder. The
descriptions are also updated.

Both `Room#event_interact` and `Room#exit` are improved (finally
removing transferring descriptions over `modifiers` variable) in respect
to the changes concerning events.