Fix event complete load bug, Add crest description
Upon finishing the fight, saving and exiting the game will load in the
armor room instead to whatever room leads to it. This commit addresses
this by adding a check for a `modifiers` array value `:game_load`.

I also added a new check in the `Room#exit` which skips saving if
`modifiers` array contains `:no_save`.

The solution relies to much on flags which I will try to remove and
improve in the future if this becomes used too much.

Also added crest description to help navigation.