Document EVERYTHING!
Well at least most of it that might be a good read. This was much harder
than I though it would be. For most of the documentation I was trying to
explain when and how should certain classes and methods be used as well
as describing how some of it works.

I spend most time on files in the lib and less time on the files in the
folders within. Especially on the room.rb, room_connecting_tools.rb,
dungeon.rb, book.rb, thesaurus.rb and satchel.rb since I feel that those
might be the most important.

Also fixed some typos and bugs here and there.