Move `DungeonTest`, `RoomTest` to integration
The above mentioned test are not (as I understand) unit tests so
I moved them there.

I severely lack in the testing compartment and I really should test
more. I'll add tests for any error or problem that occurs in the future,
since testing first is something that I just can adapt to for now.

Added description for the end room.

Added release date in CHANGELOG.md for v0.0.1